pyproject.toml
[tool.poetry]
name = "reelsbot"
version = "0.1.0"
description = ""
authors = ["Abdullakh Mursalov <a.k.mursalov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
instaloader = "^4.9.5"
python-telegram-bot = {version = "20.0a5", extras = ["webhooks"]}
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
